% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dependency.R
\name{example_dependency}
\alias{example_dependency}
\title{Dependency Example}
\usage{
example_dependency(x)
}
\arguments{
\item{x}{A numeric data series}
}
\description{
This is an exmple to show the use of dependencies
}
\examples{
df <- MASS::Boston
age <- df$age
example_dependency(age)



}
