% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_stline.R
\name{intersect_stline}
\alias{intersect_stline}
\title{Intersection for straight line}
\usage{
intersect_stline(x, y)
}
\arguments{
\item{x}{An object of the class \code{stline}}

\item{y}{An object of the class \code{stline}}
}
\value{
A numeric value of length 2, including the
\code{x} coordinate and the \code{y} coordinate of the intersecting
point
}
\description{
Intersection for straight line
}
\examples{
line_1 <- straighten(1,1,-1)
line_2 <- straighten(1,-1, 0)
intersect_stline(line_1,line_2)
}
\seealso{
\code{\link{straighten}}
}
