% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stline.R
\name{print.stline}
\alias{print.stline}
\title{Print \code{"stline"} Object}
\usage{
\method{print}{stline}(x, ... = NULL)
}
\arguments{
\item{x}{An object of class \code{"stline"}}

\item{...}{NULL; supplied for S3 method consistency}
}
\description{
Print \code{"stline"} Object
}
\examples{
line <- straighten(1, 0, 5)
print(line)

line2 <- straighten(7, -5, 0)
print(line2)

}
