% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straighten.R
\name{straighten}
\alias{straighten}
\title{Straight Line Measures}
\source{
\url{https://github.com/riazakhan94/straightline}
}
\usage{
straighten(a, b, c)
}
\arguments{
\item{a}{Parameter \eqn{a}}

\item{b}{Parameter \eqn{b}}

\item{c}{Parameter \eqn{c}}
}
\value{
A list of class \code{"stline"}, with all the input parameters
and measures described in details.
}
\description{
The function calculates slope, x intercept, y intercept
and angle with x axis of a straight line
}
\details{
If a straight line is given by \eqn{ax+by+c=0}
then slope, x intercept, y intercept and the angle made
by the line with x axis are given by \eqn{-a/b}, \eqn{-c/a},
\eqn{-c/b} and \eqn{arctan(slope)}
}
\examples{
line <- straighten(-4,4,30)
class(line)
message("The slope of the line is ", line$slope)


}
\references{
\url{https://github.com/riazakhan94/straightline}
}
\seealso{
\code{\link{hello}}
}
\author{
Jon Smith
}
